apps/W5XY-QSL-CARD-CREATOR/requirements_google.txt
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
